优化上传文件路径判断，如果没有文件夹则创建文件夹

+++ requestHandlers.js
@@ -24,18 +24,25 @@ function upload(response, request) {
   if (request.method.toLowerCase() == "post") {
     var form = new formidable.IncomingForm();
     form.uploadDir = "tmp";
-    form.parse(request, function(error, fields, files) {
-      fs.renameSync(files.upload.path, "tmp/test.png");
-      response.writeHead(200, {"Content-Type": "text/html"});
-      response.write("received image:\n\n");
-      response.write("<img src='/show' />");
-      response.end();
+    fs.exists(form.uploadDir, function(exists) {
+      if (!exists) {
+        fs.mkdir(form.uploadDir, function(err) {
+          if (!err) {
+            form.parse(request, function(error, fields, files) {
+              fs.renameSync(files.upload.path, "tmp/test.png");
+              uploadHandle(response);
+            });
+          }
+        });
+      } else {
+        form.parse(request, function(error, fields, files) {
+          fs.renameSync(files.upload.path, "tmp/test.png");
+          uploadHandle(response);
+        });
+      }
     });
   } else {
-    response.writeHead(200, {"Content-Type": "text/html"});
-    response.write("received image:\n\n");
-    response.write("<img src='/show' />");
-    response.end();
+    uploadHandle(response);
   }
 }
 
@@ -51,6 +58,13 @@ function show(response) {
       response.end();
     }
   });
+}
+
+function uploadHandle(response) {
+  response.writeHead(200, {"Content-Type": "text/html"});
+  response.write("received image:\n\n");
+  response.write("<img src='/show' />");
+  response.end();
 }
 
 exports.start = start;
